backend/requirements.txt
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
python-dotenv==1.2.2
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
google-api-python-client==2.194.0
requests==2.32.5
whitenoise==6.9.0
dj-database-url==2.3.0
gunicorn==23.0.0
qrcode==8.2
pillow==12.1.1
openpyxl==3.1.5
apscheduler==3.10.4
